Fix git cyclic dependencies
Fixes #3302
